fix: 修复 R6 分析报告 25 项缺陷 (7 Critical + 9 Important + 9 Low)
P0 Critical (7 项):
- B-7: clearForceUnlocked 改为 AtomicInteger 位标志,翻转原子化
- R-1: shutdown 后 tryLock/lock 抛 IllegalStateException,防止 lockMap 内存泄漏
- R-2: asyncExecutor rejection handler 改为 shutdown 抛 REE,业务方可降级
- B-1: tryAcquire fast-fail 分支正确计入 failedLocks
- B-5: forceUnlock 后 holderHoldCount 用 CAS 钳制到 0,不再变负
- A-1: DistributedLock SPI JavaDoc 示例补充 name 参数
- E-1: 用 System.nanoTime() 取代 Clock.millis() 计算 elapsed,Clock 回退安全

P1 Important (9 项):
- B-2: maxKeys 拒绝的 key 在 lockMap.remove 之前持久化 per-key 失败计数
- B-3: cleanupTick 借锁改用 tryLock(0, MILLISECONDS),不污染统计
- B-4/A-3: getHoldCount 改读 holderHoldCount,跨线程读一致
- A-2: tryAcquireAsync JavaDoc 重新对齐实现
- A-4: tryMultiLock 添加总 deadline 判断,超时受控
- E-2: cleanupTick 异常粒度改为 catch(Exception)
- E-3: fireXxx 异常隔离改为 catch(Throwable)
- E-4: fireContended 改为只要检测到争用就触发
- E-5: 新增 validateKey() 拒绝空字符串 key
- E-6: LockListener JavaDoc 文档化'快且非阻塞'约束

P2 Low (9 项):
- A-5: 新增 MaxKeysExceededException(继承 IllegalStateException)
- A-6: lockInterruptibly JavaDoc 文档化 InterruptedException
- A-7: registerListener(null) 抛 NPE,与 registerDistributedLock 一致
- E-8: JavaDoc 全面扫尾
- E-9: holderHoldCount 改 AtomicLong,防溢出
- L-1/L-2: CHMRLockConfig.Builder setter 校验 null/负数
- L-3: KeyedReadWriteLock JavaDoc 文档化 IllegalMonitorStateException
- L-9: exportMetrics shutdown 后仍可读,perKey 为空

测试:
- 新增 6 轮专项测试 (Round1..Round6 + BugDetective) 共 110+ 用例
- 修改 6 个现有测试以适配 R6 行为变更
- 全量 280 测试通过

+++ src/main/java/cn/wubo/lock/KeyStatistics.java
@@ -2,7 +2,10 @@ package cn.wubo.lock;
 
 /**
  * 单个 key 的统计指标。
- * 由 CHMRLock.getStatistics(String key) 返回，仅在 CHMRLockConfig.enablePerKeyMetrics=true 时收集。
+ * 由 CHMRLock.getStatistics(String key) 返回,仅在 CHMRLockConfig.enablePerKeyMetrics=true 时收集。
+ *
+ * <p>R6 修复(E-9):{@link #currentHoldCount()} 由 {@code int} 改为 {@code long},
+ * 与 {@link LockEntry#getHolderHoldCount()} 一致,避免极端重入深度下溢出。</p>
  *
  * @see CHMRLock#getStatistics(String)
  * @see CHMRLock#getAllStatistics()
@@ -18,7 +21,7 @@ public record KeyStatistics(
         /** 累计的获取失败次数(超时 / maxKeys / 中断)。 */
         long failedCount,
         /**
-         * 累计等待时间(纳秒)。通过 System.nanoTime() 测量，精度为纳秒级。
+         * 累计等待时间(纳秒)。通过 System.nanoTime() 测量,精度为纳秒级。
          * 注意:包含成功与失败两种获取的总等待时间。
          */
         long totalWaitNanos,
@@ -27,7 +30,7 @@ public record KeyStatistics(
         /** 最近一次成功释放锁的 epoch 毫秒时间戳;0 表示从未释放。 */
         long lastReleaseEpochMs,
         /** 当前线程对该 key 的重入深度(0 表示未持有)。 */
-        int currentHoldCount,
+        long currentHoldCount,
         /** 当前持有该 key 的线程 id;-1 表示未被持有。 */
         long currentHolderThreadId
 ) {
